requirements.txt
torch~=2.0.0+cu117
dgl-cu111~=0.8.2
numpy~=1.22.2
scipy~=1.7.1
tqdm~=4.62.3